requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.29.0
pinecone-client==3.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
langchain==0.0.340
unstructured==0.11.6
sentence-transformers==2.2.2
google-generativeai==0.3.2
python-multipart==0.0.6
pypdf==3.17.1
python-docx==1.1.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.25.2
requests==2.31.0
aiofiles==23.2.0
pydantic==2.5.2
faiss-cpu==1.7.4
plotly==5.17.0
